feat: add IPConfig state logs/metrics to Request/Release handlers
