+++ cns/restserver/internalapi.go
@@ -615,7 +615,7 @@ func (service *HTTPRestService) CreateOrUpdateNetworkContainerInternal(req *cns.
 	// If the NC was created successfully, log NC snapshot.
 	if returnCode == 0 {
 		logNCSnapshot(*req)
-		publishIPStateMetrics(service.buildIPState())
+		service.publishIPStateMetrics()
 	} else {
 		logger.Errorf(returnMessage)
 	}
